Moved ConfigOverrides test in a separate file. Moved helpers to TestBase and TestHelpers. Refactored some helper code to make it extensible and reusable.
